Samuel's aka Twin Brother mod
Hey dudes, I want to make a long fnf mod that starts with my creation "Samuel'. Samuel is a 23 year old and 10 feet tall man who love to rap battle, he also got powers because of the collision that happen with the demons and the angels. Samuel is unstoppable, got millions of power, he's handsome, he's very handsome that he attract a lot of hotties that Boyfriend rap battled, If you want to join all of the fun, come and join to help me settle this long mod that everyone wants to play